# Conflicts:
#   generic-service-post/src/test/java/nl/tudelft/sem/genericservicepost/services/GenericPostServiceTest.java

+++ generic-service-post/src/test/java/nl/tudelft/sem/genericservicepost/services/GenericPostServiceTest.java
@@ -34,8 +34,8 @@ public class GenericPostServiceTest {
         genericPost = new GenericPost();
         genericPost1 = new GenericPost();
 
-        genericPost.setId((long) 0);
-        genericPost1.setId((long) 1);
+        genericPost.setId(0L);
+        genericPost1.setId(1L);
 
         genericPost.setAuthor("The Rock");
         genericPost1.setAuthor("The Wok");
